# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Core backend service for BookEat.\n\nEvery response is wrapped in a JSON envelope — {\"data\": ...} on\nsuccess, {\"error\": \"...\"} on failure — except /.well-known/jwks.json,\nwhich returns a raw JWKS document.",
        "title": "BookEat backend-core API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/api/v1/admin/my-restaurants": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Restaurants where the caller is owner/manager/hostess, with their\nrole at each. A superadmin gets every restaurant (role \"admin\").",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "admin"
                ],
                "summary": "List restaurants I am staff of",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/devices/push-tokens": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "push"
                ],
                "summary": "Register my device for push notifications",
                "parameters": [
                    {
                        "description": "Provider push token + platform",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/devicetokens.registerRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "422": {
                        "description": "invalid body",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "push"
                ],
                "summary": "Unregister my device from push notifications",
                "parameters": [
                    {
                        "description": "The token to silence",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/devicetokens.unregisterRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "401": {
                        "description": "unauthorized",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "422": {
                        "description": "invalid body",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
        "/api/v1/restaurants/{id}/map": {
            "get": {
                "description": "Returns a PNG map preview of the restaurant, rendered server-side so the map provider key never reaches the client. Size and zoom are whitelisted. Failures return a JSON envelope whose \"code\" tells them apart: map_no_coordinates (404), map_not_configured (503), map_provider_rate_limited (503), map_provider_unavailable (503).",
                "produces": [
                    "image/png"
                ],
                "tags": [
                    "restaurants"
                ],
                "summary": "Restaurant map preview (server-side proxy)",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Restaurant id",
                        "name": "id",
                        "in": "path"
                    },
                    {
                        "enum": [
                            "card",
                            "detail",
                            "wide"
                        ],
                        "type": "string",
                        "default": "detail",
                        "description": "Image preset",
                        "name": "size",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "1 or 2 (HiDPI)",
                        "name": "scale",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 16,
                        "description": "14..18",
                        "name": "zoom",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "PNG image",
                        "schema": {
                            "type": "file"
                        }
                    },
                    "404": {
                        "description": "unknown restaurant or no coordinates",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "422": {
                        "description": "size/scale/zoom outside the whitelist",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    },
                    "503": {
                        "description": "provider not configured or unavailable",
                        "schema": {
                            "$ref": "#/definitions/response.Envelope"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "devicetokens.registerRequest": {
            "type": "object",
            "required": [
                "platform",
                "token"
            ],
            "properties": {
                "platform": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "devicetokens.unregisterRequest": {
            "type": "object",
            "required": [
                "token"
            ],
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
...
        "response.Envelope": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "data": {},
                "error": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Access token from POST /api/v1/auth/login (or signup / otp/verify /\nrefresh), sent as: Authorization: Bearer \u003caccess_token\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}